pyproject.toml
[project]
name = "flask-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.0",
]
